Explain the output of machine learning models with more accurately estimated Shapley values

TODO: Add a more detailed description

explain(x, explainer, approach, prediction_zero, ...)

# S3 method for empirical
explain(x, explainer, approach, prediction_zero,
  type = "fixed_sigma", fixed_sigma_vec = 0.1,
  AICc_no_samp_per_optim = 1000, AIC_optim_max_eval = 20,
  AIC_optim_startval = 0.1, w_threshold = 0.95, ...)

# S3 method for gaussian
explain(x, explainer, approach, prediction_zero,
  mu = NULL, cov_mat = NULL, ...)

# S3 method for copula
explain(x, explainer, approach, prediction_zero, ...)

# S3 method for combined
explain(x, explainer, approach, prediction_zero,
  mu = NULL, cov_mat = NULL, ...)

Arguments

x
A matrix or data.frame. Contains the the features, whose predictions ought to be explained (test data).

explainer
An explainer object to use for explaining the observations. See shapr.

approach
Character vector of length 1 or n_features. n_features equals the total number of features in the model. All elements should either be "gaussian", "copula" or "empirical". See details for more information.

prediction_zero
The prediction value for unseen data, typically equal to the mean of the response.

...
Additional arguments passed to prepare_data

type
Character. Should be equal to either "independence", "fixed_sigma", "AICc_each_k" or "AICc_full".

fixed_sigma_vec
Vector or numeric. Only applicable when approach='empirical' and type='fixed_sigma'. The bandwidth to use. Default value 0.1

AICc_no_samp_per_optim
Positive integer. Only applicable when approach='empirical' and type='AICc_each_k' or type='AICc_full'. Number of samples to consider in AICc optimization.

AIC_optim_max_eval
Positive integer. Only applicable when approach='empirical' and type='AICc_each_k' or type='AICc_full'. Numeric. Maximum value when optimizing the AICc.

AIC_optim_startval
Numeric. Only applicable when approach='empirical' and type='AICc_each_k' or type='AICc_full'. Starting value when optimizing the AICc.

w_threshold
Positive integer between 0 and 1.

mu
Numeric vector. (Optional) Containing the mean of the data generating distribution. If NULL the expected values are estimated from the data. Note that this is only used when approach = "gaussian".

cov_mat
Numeric matrix. (Optional) Containing the covariance matrix of the data generating distribution. NULL means it is estimated from the data if needed (in the Gaussian approach).

Value

data.frame. Contains the estimated Shapley values for the test data. Note that the dimensions of the data.frame equals n x (p+1), where n equals the number of test observations, and p equals the total number of features.

Details

TODO: Add information about approach. TODO: Some additional details about the returned object