Sample conditional variables using the Gaussian copula approach

Sample conditional variables using the Gaussian copula approach

sample_copula(
  index_given,
  n_samples,
  mu,
  cov_mat,
  m,
  x_explain_gaussian,
  x_train,
  x_explain
)

Arguments

index_given

Integer vector. The indices of the features to condition upon. Note that min(index_given) >= 1 and max(index_given) <= m.

m

Positive integer. The total number of features.

x_explain_gaussian

Numeric matrix. Contains the observation whose predictions ought to be explained (test data), after quantile-transforming them to standard Gaussian variables.

x_explain

Numeric matrix. Contains the features of the observation whose predictions ought to be explained (test data).

Value

data.table

Author

Martin Jullum