Sample ctree variables from a given conditional inference tree

Sample ctree variables from a given conditional inference tree

sample_ctree(tree, n_samples, x_test, x_train, p, sample)

Arguments

tree

List. Contains tree which is an object of type ctree built from the party package. Also contains given_ind, the features to condition upon.

n_samples

Numeric. Indicates how many samples to use for MCMC.

x_test

Matrix, data.frame or data.table with the features of the observation whose predictions ought to be explained (test data). Dimension 1xp or px1.

x_train

Matrix, data.frame or data.table with training data.

p

Positive integer. The number of features.

sample

Boolean. True indicates that the method samples from the terminal node of the tree whereas False indicates that the method takes all the observations if it is less than n_samples.

Value

data.table with n_samples (conditional) Gaussian samples

Author

Annabelle Redelmeier

Examples

if (requireNamespace("MASS", quietly = TRUE) & requireNamespace("party", quietly = TRUE)) {
  m <- 10
  n <- 40
  n_samples <- 50
  mu <- rep(1, m)
  cov_mat <- cov(matrix(rnorm(n * m), n, m))
  x_train <- data.table::data.table(MASS::mvrnorm(n, mu, cov_mat))
  x_test <- MASS::mvrnorm(1, mu, cov_mat)
  x_test_dt <- data.table::setDT(as.list(x_test))
  given_ind <- c(4, 7)
  dependent_ind <- (1:dim(x_train)[2])[-given_ind]
  x <- x_train[, given_ind, with = FALSE]
  y <- x_train[, dependent_ind, with = FALSE]
  df <- data.table::data.table(cbind(y, x))
  colnames(df) <- c(paste0("Y", 1:ncol(y)), paste0("V", given_ind))
  ynam <- paste0("Y", 1:ncol(y))
  fmla <- as.formula(paste(paste(ynam, collapse = "+"), "~ ."))
  datact <- party::ctree(fmla, data = df, controls = party::ctree_control(
    minbucket = 7,
    mincriterion = 0.95
  ))
  tree <- list(tree = datact, given_ind = given_ind, dependent_ind = dependent_ind)
  shapr:::sample_ctree(
    tree = tree, n_samples = n_samples, x_test = x_test_dt, x_train = x_train,
    p = length(x_test), sample = TRUE
  )
}
#>             V1           V2         V3        V4          V5          V6
#>  1:  0.5318019  2.566985356  1.4622643 0.9579884 -0.01487250  1.89966684
#>  2: -0.6753265 -0.006084004  0.3919006 0.9579884  2.86918351  1.00103451
#>  3: -1.9842188 -1.184952561  1.7083584 0.9579884  1.78333408  0.51034367
#>  4:  1.7240279  1.692814638 -0.1218454 0.9579884  2.16614656 -0.04644525
#>  5: -0.2822999  1.671308872 -0.1488490 0.9579884  0.48609593  2.61681085
#>  6:  0.2110467 -0.109578148 -0.6594373 0.9579884  2.26318268  0.87796212
#>  7:  0.8819050  1.293741779  1.6465323 0.9579884  0.01075943  2.26670312
#>  8:  1.6007318  2.848824676  0.4060111 0.9579884 -0.56852696  0.77234937
#>  9: -0.1408801  1.403236567  2.0735028 0.9579884  0.09757056  2.04294474
#> 10:  1.9282250  2.368336916 -0.8623474 0.9579884  1.51239676  2.13225054
#> 11:  3.0337297  0.846083253  0.1804255 0.9579884 -0.19794671 -1.22759989
#> 12:  0.8220553 -0.204361688  0.4903171 0.9579884  2.49291305  0.83968665
#> 13:  0.5926165  0.527648132  0.7006535 0.9579884  0.86916937  1.31474460
#> 14:  3.0337297  0.846083253  0.1804255 0.9579884 -0.19794671 -1.22759989
#> 15:  2.1738558  2.434909510  1.3530631 0.9579884  1.83924488  2.79757684
#> 16:  1.1096676  0.943049603  0.5116621 0.9579884  0.57359803  0.59251442
#> 17:  1.9282250  2.368336916 -0.8623474 0.9579884  1.51239676  2.13225054
#> 18:  1.1193928  2.601723315  0.2262768 0.9579884  1.07115712 -0.71776702
#> 19:  0.8874391  1.022233931  0.8958387 0.9579884  1.56877981  0.79693864
#> 20:  1.1193928  2.601723315  0.2262768 0.9579884  1.07115712 -0.71776702
#> 21:  1.0080789  1.587664814  1.5163513 0.9579884  1.32909125  1.04219594
#> 22: -1.9842188 -1.184952561  1.7083584 0.9579884  1.78333408  0.51034367
#> 23:  0.7832831  1.717663703  0.4505747 0.9579884  0.48987205 -0.40334074
#> 24:  1.6007318  2.848824676  0.4060111 0.9579884 -0.56852696  0.77234937
#> 25: -0.8620888 -0.529703197  0.6969587 0.9579884  0.26405572  0.22038670
#> 26: -0.2822999  1.671308872 -0.1488490 0.9579884  0.48609593  2.61681085
#> 27: -0.5898974  1.858487852  1.4917106 0.9579884  0.96437093  2.61931103
#> 28:  0.6382765  1.351887986  1.1714350 0.9579884 -0.14051040  2.93074678
#> 29:  1.7240279  1.692814638 -0.1218454 0.9579884  2.16614656 -0.04644525
#> 30:  1.7240279  1.692814638 -0.1218454 0.9579884  2.16614656 -0.04644525
#> 31:  3.0337297  0.846083253  0.1804255 0.9579884 -0.19794671 -1.22759989
#> 32:  1.2497974  1.447496126  1.4151783 0.9579884  0.31507838 -1.71024026
#> 33:  1.7557340  1.882381653  1.0647390 0.9579884 -0.36810177  0.38640779
#> 34:  2.5349542  3.466574547  0.8616325 0.9579884 -1.02973213  2.57842835
#> 35:  2.0825333 -0.361695421 -0.4992902 0.9579884  2.16855195  0.76439770
#> 36:  1.1096676  0.943049603  0.5116621 0.9579884  0.57359803  0.59251442
#> 37: -0.5898974  1.858487852  1.4917106 0.9579884  0.96437093  2.61931103
#> 38:  1.4683589  0.627522228  1.3341940 0.9579884  1.64545663  1.92825113
#> 39:  1.0080789  1.587664814  1.5163513 0.9579884  1.32909125  1.04219594
#> 40:  2.0825333 -0.361695421 -0.4992902 0.9579884  2.16855195  0.76439770
#> 41:  1.7557340  1.882381653  1.0647390 0.9579884 -0.36810177  0.38640779
#> 42: -0.8620888 -0.529703197  0.6969587 0.9579884  0.26405572  0.22038670
#> 43:  1.1309652  1.102918341  0.5525010 0.9579884  0.43172165  0.14200660
#> 44:  1.0080789  1.587664814  1.5163513 0.9579884  1.32909125  1.04219594
#> 45:  0.8819050  1.293741779  1.6465323 0.9579884  0.01075943  2.26670312
#> 46: -1.9842188 -1.184952561  1.7083584 0.9579884  1.78333408  0.51034367
#> 47:  2.5349542  3.466574547  0.8616325 0.9579884 -1.02973213  2.57842835
#> 48:  0.6382765  1.351887986  1.1714350 0.9579884 -0.14051040  2.93074678
#> 49:  2.5349542  3.466574547  0.8616325 0.9579884 -1.02973213  2.57842835
#> 50:  0.8220553 -0.204361688  0.4903171 0.9579884  2.49291305  0.83968665
#>             V1           V2         V3        V4          V5          V6
#>           V7          V8          V9          V10
#>  1: 1.404042  2.09852537  1.21070033  1.448713741
#>  2: 1.404042 -0.24018270  2.18841582 -0.081032169
#>  3: 1.404042 -1.08361448  2.46703199 -2.146649631
#>  4: 1.404042  1.01507326 -0.39720128  0.827752458
#>  5: 1.404042  1.86726113  1.89190583  1.001560802
#>  6: 1.404042  0.80893194  1.42759234  1.411807833
#>  7: 1.404042  2.17292231  1.79025550  0.410034918
#>  8: 1.404042  2.13472930  1.85936105  0.694472259
#>  9: 1.404042  0.58315305  2.43962828  0.887494936
#> 10: 1.404042  0.50533299  1.13138297  2.102415999
#> 11: 1.404042  1.54902910  1.53572191  0.980232875
#> 12: 1.404042  2.08695407  0.79776331  1.838787162
#> 13: 1.404042  1.07547916  1.24726747  2.602663796
#> 14: 1.404042  1.54902910  1.53572191  0.980232875
#> 15: 1.404042  2.42184148  2.11615133 -1.398609286
#> 16: 1.404042  0.83812580  0.58524414  1.451563525
#> 17: 1.404042  0.50533299  1.13138297  2.102415999
#> 18: 1.404042  1.58452987  1.24067311  1.353839967
#> 19: 1.404042 -0.30197895  1.16574613  0.691567636
#> 20: 1.404042  1.58452987  1.24067311  1.353839967
#> 21: 1.404042 -0.03895497  0.23725481  3.734943583
#> 22: 1.404042 -1.08361448  2.46703199 -2.146649631
#> 23: 1.404042  1.03479391  1.44544799  0.007798773
#> 24: 1.404042  2.13472930  1.85936105  0.694472259
#> 25: 1.404042  0.70268807  2.33097646  1.003209127
#> 26: 1.404042  1.86726113  1.89190583  1.001560802
#> 27: 1.404042  1.35238124  2.02378861 -0.683973297
#> 28: 1.404042  1.58899936  2.88815547  1.155341436
#> 29: 1.404042  1.01507326 -0.39720128  0.827752458
#> 30: 1.404042  1.01507326 -0.39720128  0.827752458
#> 31: 1.404042  1.54902910  1.53572191  0.980232875
#> 32: 1.404042 -0.02261701  2.26562204  2.461995070
#> 33: 1.404042  0.72445327  0.09948112  2.199359327
#> 34: 1.404042  2.05476876  1.65713861  0.086251539
#> 35: 1.404042  0.98694474 -1.86060831  2.396405736
#> 36: 1.404042  0.83812580  0.58524414  1.451563525
#> 37: 1.404042  1.35238124  2.02378861 -0.683973297
#> 38: 1.404042  2.70937920 -0.26380695  0.601677581
#> 39: 1.404042 -0.03895497  0.23725481  3.734943583
#> 40: 1.404042  0.98694474 -1.86060831  2.396405736
#> 41: 1.404042  0.72445327  0.09948112  2.199359327
#> 42: 1.404042  0.70268807  2.33097646  1.003209127
#> 43: 1.404042  2.02568376  0.96788859  0.329282571
#> 44: 1.404042 -0.03895497  0.23725481  3.734943583
#> 45: 1.404042  2.17292231  1.79025550  0.410034918
#> 46: 1.404042 -1.08361448  2.46703199 -2.146649631
#> 47: 1.404042  2.05476876  1.65713861  0.086251539
#> 48: 1.404042  1.58899936  2.88815547  1.155341436
#> 49: 1.404042  2.05476876  1.65713861  0.086251539
#> 50: 1.404042  2.08695407  0.79776331  1.838787162
#>           V7          V8          V9          V10