A torch::nn_module() Representing a skip connection

Skip connection over the sequence of layers in the constructor. The module passes input data sequentially through these layers and then adds original data to the result.

Usage

skip_connection(...)

Arguments

...

network modules such as, e.g., torch::nn_linear(), torch::nn_relu(), and memory_layer() objects. See vaeac() for more information.

Author

Lars Henry Berge Olsen