Function that checks the specified masking scheme

Function that checks the specified masking scheme

vaeac_check_mask_gen(mask_gen_coalitions, mask_gen_coalitions_prob, x_train)

Arguments

mask_gen_coalitions

Matrix (default is NULL). Matrix containing the coalitions that the vaeac model will be trained on, see specified_masks_mask_generator(). This parameter is used internally in shapr when we only consider a subset of coalitions/combinations, i.e., when n_combinations \(< 2^{n_{\text{features}}}\), and for group Shapley, i.e., when group is specified in explain().

mask_gen_coalitions_prob

Numeric array (default is NULL). Array of length equal to the height of mask_gen_coalitions containing the probabilities of sampling the corresponding coalitions in mask_gen_coalitions.

x_train

A data.table containing the training data. Categorical data must have class names \(1,2,\dots,K\).

Value

The function does not return anything.

Author

Lars Henry Berge Olsen