Function to create the optimizer used to train vaeac

Only torch::optim_adam() is currently supported. But it is easy to add an additional option later.

Usage

vaeac_get_optimizer(vaeac_model, lr, optimizer_name = "adam")

Arguments

vaeac_model

A vaeac model created using vaeac().

lr

Positive numeric (default is 0.001). The learning rate used in the torch::optim_adam() optimizer.

optimizer_name

String containing the name of the torch::optimizer() to use.

Value

A torch::optim_adam() optimizer connected to the parameters of the vaeac_model.

Author

Lars Henry Berge Olsen