Generate permutations of training data using test observations

Generate permutations of training data using test observations

observation_impute(
  W_kernel,
  S,
  x_train,
  x_test,
  w_threshold = 0.7,
  n_samples = 1000
)

Arguments

W_kernel
Numeric matrix. Contains all nonscaled weights between training and test observations for all feature combinations. The dimension equals n_train x m.

S
Integer matrix of dimension n_combinations x m, where n_combinations and m equals the total number of sampled/non-sampled feature combinations and the total number of unique features, respectively. Note that m = ncol(x_train).

x_train
Numeric matrix

x_test
Numeric matrix

w_threshold
Numeric vector of length 1, where w_threshold > 0 and w_threshold <= 1. If w_threshold = .8 we will choose the K samples with the largest weight so that the sum of the weights accounts for 80% of the total weight.

Value

data.table

Author

Nikolai Sellereite

Examples

# Setup
n <- 20 # Sample size of training data
m <- 2 # Number of features
sigma <- cov(matrix(MASS::mvrnorm(m * n, 0, 1), nrow = n))

# Create training- and test data
x_train <- as.matrix(MASS::mvrnorm(n, mu = rep(0, m), Sigma = sigma), ncol = m)
x_test <- t(as.matrix(MASS::mvrnorm(1, mu = rep(0, m), sigma)))
colnames(x_train) <- colnames(x_test) <- paste0("X", seq(m))

# Binary matrix which represents the feature combinations
S <- matrix(c(1, 0, 0, 1), nrow = m)

# Kernel matrix
W_kernel <- matrix(rnorm(n * ncol(S), mean = 1 / n, sd = 1 / n^2), nrow = n)

# Generate permutations of training data using test observations
r <- shapr:::observation_impute(W_kernel, S, x_train, x_test)
str(r)

#> Classes ‘data.table’ and 'data.frame':	28 obs. of  4 variables:
#>  $ X1            : num  1.98 1.98 1.98 1.98 1.98 ...
#>  $ X2            : num  2.91 0.175 1.072 -0.203 1.006 ...
#>  $ id_combination: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ w             : num  0.049 0.0492 0.0494 0.0499 0.0502 ...
#>  - attr(*, ".internal.selfref")=<externalptr>