Sample ctree variables from a given conditional inference tree

Sample ctree variables from a given conditional inference tree

sample_ctree(tree, n_samples, x_test, x_train, p, sample)

Arguments

tree
List. Contains tree which is an object of type ctree built from the party package. Also contains given_ind, the features to condition upon.

n_samples
Numeric. Indicates how many samples to use for MCMC.

x_test
Matrix, data.frame or data.table with the features of the observation whose predictions ought to be explained (test data). Dimension 1xp or px1.

x_train
Matrix, data.frame or data.table with training data.

p
Positive integer. The number of features.

sample
Boolean. True indicates that the method samples from the terminal node of the tree whereas False indicates that the method takes all the observations if it is less than n_samples.

Value

data.table with n_samples (conditional) Gaussian samples

Author

Annabelle Redelmeier

Examples

if (requireNamespace("MASS", quietly = TRUE) & requireNamespace("party", quietly = TRUE)) {
  m <- 10
  n <- 40
  n_samples <- 50
  mu <- rep(1, m)
  cov_mat <- cov(matrix(rnorm(n * m), n, m))
  x_train <- data.table::data.table(MASS::mvrnorm(n, mu, cov_mat))
  x_test <- MASS::mvrnorm(1, mu, cov_mat)
  x_test_dt <- data.table::setDT(as.list(x_test))
  given_ind <- c(4, 7)
  dependent_ind <- (1:dim(x_train)[2])[-given_ind]
  x <- x_train[, given_ind, with = FALSE]
  y <- x_train[, dependent_ind, with = FALSE]
  df <- data.table::data.table(cbind(y, x))
  colnames(df) <- c(paste0("Y", 1:ncol(y)), paste0("V", given_ind))
  ynam <- paste0("Y", 1:ncol(y))
  fmla <- as.formula(paste(paste(ynam, collapse = "+"), "~ ."))
  datact <- party::ctree(fmla, data = df, controls = party::ctree_control(
    minbucket = 7,
    mincriterion = 0.95
  ))
  tree <- list(tree = datact, given_ind = given_ind, dependent_ind = dependent_ind)
  shapr:::sample_ctree(
    tree = tree, n_samples = n_samples, x_test = x_test_dt, x_train = x_train,
    p = length(x_test), sample = TRUE
  )
}

#>               V1         V2         V3        V4          V5          V6     V7
#>  1: -1.093876376  1.1063321  1.4138340 0.5981535  0.09245859  0.92925925 1.8679
#>  2:  2.030356116  2.4249408  0.4725051 0.5981535  2.69054801  2.61612343 1.8679
#>  3:  2.528226051  2.8694275  1.8427877 0.5981535  1.48541189  3.20393373 1.8679
#>  4:  0.322734528  0.4337686 -0.1635911 0.5981535  2.25866323 -0.88291195 1.8679
#>  5: -0.008143035  1.9176357 -0.1406817 0.5981535  0.46799545  2.78046187 1.8679
#>  6:  2.098754799  1.5865060 -0.6032010 0.5981535  2.13855177  2.00478184 1.8679
#>  7:  0.684481179  1.1163587  1.6406509 0.5981535  0.02379382  2.14885596 1.8679
#>  8:  0.455246784  1.8196195  0.3718862 0.5981535 -0.49289936  0.08858102 1.8679
#>  9:  0.270840134  1.7731625  2.0857683 0.5981535  0.07038783  2.28871073 1.8679
#> 10:  0.614962079  1.1883852 -0.9014706 0.5981535  1.59910146  1.34833136 1.8679
#> 11:  2.682153317  0.5301959  0.1699518 0.5981535 -0.17473481 -1.43746455 1.8679
#> 12:  1.186301139  0.1229090  0.5011683 0.5981535  2.46886469  1.05711401 1.8679
#> 13:  0.648856391  0.5781791  0.7023290 0.5981535  0.86545628  1.34831560 1.8679
#> 14:  2.682153317  0.5301959  0.1699518 0.5981535 -0.17473481 -1.43746455 1.8679
#> 15:  0.862979803  1.2571024  1.3140111 0.5981535  1.92579199  2.01508250 1.8679
#> 16:  0.680810709  0.5577266  0.4988861 0.5981535  0.60191217  0.33651910 1.8679
#> 17:  0.614962079  1.1883852 -0.9014706 0.5981535  1.59910146  1.34833136 1.8679
#> 18:  0.685938495  2.2122696  0.2133639 0.5981535  1.09977478 -0.97650664 1.8679
#> 19:  1.330501055  1.4203200  0.9090379 0.5981535  1.53952782  1.06141329 1.8679
#> 20:  0.685938495  2.2122696  0.2133639 0.5981535  1.09977478 -0.97650664 1.8679
#> 21:  0.507870511  1.1382333  1.5014497 0.5981535  1.36211618  0.74360919 1.8679
#> 22:  2.528226051  2.8694275  1.8427877 0.5981535  1.48541189  3.20393373 1.8679
#> 23:  0.796656826  1.7296799  0.4509731 0.5981535  0.48898909 -0.39535763 1.8679
#> 24:  0.455246784  1.8196195  0.3718862 0.5981535 -0.49289936  0.08858102 1.8679
#> 25:  1.921632138  1.9714381  0.7798880 0.5981535  0.08026794  1.88205863 1.8679
#> 26: -0.008143035  1.9176357 -0.1406817 0.5981535  0.46799545  2.78046187 1.8679
#> 27:  0.019619358  2.4061316  1.5098686 0.5981535  0.92412921  2.98314666 1.8679
#> 28:  0.769857903  1.4701124  1.1753549 0.5981535 -0.14919772  3.00929100 1.8679
#> 29:  0.322734528  0.4337686 -0.1635911 0.5981535  2.25866323 -0.88291195 1.8679
#> 30:  0.322734528  0.4337686 -0.1635911 0.5981535  2.25866323 -0.88291195 1.8679
#> 31:  2.682153317  0.5301959  0.1699518 0.5981535 -0.17473481 -1.43746455 1.8679
#> 32:  2.622883797  2.6811986  1.4560837 0.5981535  0.22442400 -0.89061097 1.8679
#> 33:  0.464186524  0.7219411  1.0262627 0.5981535 -0.28283078 -0.38454886 1.8679
#> 34: -0.459505519  0.7760868  0.7724252 0.5981535 -0.83203089  0.79096121 1.8679
#> 35:  0.724381163 -1.5819796 -0.5397507 0.5981535  2.25822034 -0.04631696 1.8679
#> 36:  0.680810709  0.5577266  0.4988861 0.5981535  0.60191217  0.33651910 1.8679
#> 37:  0.019619358  2.4061316  1.5098686 0.5981535  0.92412921  2.98314666 1.8679
#> 38:  0.419537259 -0.3148320  1.3029488 0.5981535  1.71470229  1.30218355 1.8679
#> 39:  0.507870511  1.1382333  1.5014497 0.5981535  1.36211618  0.74360919 1.8679
#> 40:  0.724381163 -1.5819796 -0.5397507 0.5981535  2.25822034 -0.04631696 1.8679
#> 41:  0.464186524  0.7219411  1.0262627 0.5981535 -0.28283078 -0.38454886 1.8679
#> 42:  1.921632138  1.9714381  0.7798880 0.5981535  0.08026794  1.88205863 1.8679
#> 43:  0.835355233  0.8373162  0.5436946 0.5981535  0.45123851 -0.03445028 1.8679
#> 44:  0.507870511  1.1382333  1.5014497 0.5981535  1.36211618  0.74360919 1.8679
#> 45:  0.684481179  1.1163587  1.6406509 0.5981535  0.02379382  2.14885596 1.8679
#> 46:  2.528226051  2.8694275  1.8427877 0.5981535  1.48541189  3.20393373 1.8679
#> 47: -0.459505519  0.7760868  0.7724252 0.5981535 -0.83203089  0.79096121 1.8679
#> 48:  0.769857903  1.4701124  1.1753549 0.5981535 -0.14919772  3.00929100 1.8679
#> 49: -0.459505519  0.7760868  0.7724252 0.5981535 -0.83203089  0.79096121 1.8679
#> 50:  1.186301139  0.1229090  0.5011683 0.5981535  2.46886469  1.05711401 1.8679
#>               V1         V2         V3        V4          V5          V6     V7
#>               V8         V9         V10
#>  1:  1.008145864  2.7870917  1.09652567
#>  2:  1.574580356 -0.4352366  0.50512879
#>  3:  1.942985270 -1.9086057 -1.16907043
#>  4:  0.075193832  0.9616080  0.52417529
#>  5:  2.051144400  1.6260608  1.06095434
#>  6:  2.075060854 -0.4028847  1.82076224
#>  7:  2.040505649  1.9816939  0.36726489
#>  8:  1.366426383  2.9701175  0.44631359
#>  9:  0.859303195  2.0403905  0.97669029
#> 10: -0.375502441  2.4048308  1.81790980
#> 11:  1.313218779  1.8766393  0.90406712
#> 12:  2.331262071  0.4445606  1.91769764
#> 13:  1.113200562  1.1927326  2.61484765
#> 14:  1.313218779  1.8766393  0.90406712
#> 15:  1.542607029  3.3872845 -1.68259838
#> 16:  0.550481673  1.0010991  1.35865566
#> 17: -0.375502441  2.4048308  1.81790980
#> 18:  1.293802163  1.6609861  1.25993612
#> 19: -0.004807187  0.7361168  0.78755289
#> 20:  1.293802163  1.6609861  1.25993612
#> 21: -0.374456127  0.7222980  3.62657810
#> 22:  1.942985270 -1.9086057 -1.16907043
#> 23:  1.043763982  1.4324797  0.01069607
#> 24:  1.366426383  2.9701175  0.44631359
#> 25:  2.569793200 -0.3683483  1.60627636
#> 26:  2.051144400  1.6260608  1.06095434
#> 27:  1.761198017  1.4327511 -0.55192717
#> 28:  1.677254022  2.7605633  1.18384733
#> 29:  0.075193832  0.9616080  0.52417529
#> 30:  0.075193832  0.9616080  0.52417529
#> 31:  1.313218779  1.8766393  0.90406712
#> 32:  0.898343400  0.9341645  2.75946148
#> 33: -0.141817089  1.3518718  1.91955758
#> 34:  0.046316306  4.5608130 -0.56247030
#> 35:  0.076001085 -0.5436323  2.10217470
#> 36:  0.550481673  1.0010991  1.35865566
#> 37:  1.761198017  1.4327511 -0.55192717
#> 38:  2.005910617  0.7532167  0.37446014
#> 39: -0.374456127  0.7222980  3.62657810
#> 40:  0.076001085 -0.5436323  2.10217470
#> 41: -0.141817089  1.3518718  1.91955758
#> 42:  2.569793200 -0.3683483  1.60627636
#> 43:  1.827411447  1.2545363  0.26524143
#> 44: -0.374456127  0.7222980  3.62657810
#> 45:  2.040505649  1.9816939  0.36726489
#> 46:  1.942985270 -1.9086057 -1.16907043
#> 47:  0.046316306  4.5608130 -0.56247030
#> 48:  1.677254022  2.7605633  1.18384733
#> 49:  0.046316306  4.5608130 -0.56247030
#> 50:  2.331262071  0.4445606  1.91769764
#>               V8         V9         V10