Continue to Train the vaeac Model

Function that loads a previously trained vaeac model and continue the training, either on new data or on the same dataset as it was trained on before. If we are given a new dataset, then we assume that new dataset has the same distribution and one_hot_max_sizes as the original dataset.

vaeac_train_model_continue(
  explanation,
  epochs_new,
  lr_new = NULL,
  x_train = NULL,
  save_data = FALSE,
  verbose = 0,
  seed = 1
)

Arguments

explanation

A explain() object and vaeac must be the used approach.

epochs_new

Positive integer. The number of extra epochs to conduct.

lr_new

Positive numeric. If we are to overwrite the old learning rate in the adam optimizer.

x_train

A data.table containing the training data. Categorical data must have class names \(1,2,\dots,K\).

save_data

Logical (default is FALSE). If TRUE, then the data is stored together with the model. Useful if one are to continue to train the model later using vaeac_train_model_continue().

verbose

Boolean. An integer specifying the level of verbosity. Use 0 (default) for no verbosity, 1 for low verbose, and 2 for high verbose.

seed

Positive integer (default is 1). Seed for reproducibility. Specifies the seed before any randomness based code is being run.

Value

A list containing the training/validation errors and paths to where the vaeac models are saved on the disk.

Author

Lars Henry Berge Olsen